Move common logic for report generation into its own module. (#362)
No code change, just straight copy paste.

Moving the common classes into their own module was raised in #343.
